requirements.txt
pocketsphinx==0.1.15
PyYAML==3.13
retrying==1.3.3
six==1.12.0
tqdm==4.31.1
